requirements.txt
amqp==5.1.1
anyio==3.6.2
APScheduler==3.9.1
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
cachetools==5.2.0
celery==5.2.7
celery-once==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
contourpy==1.0.7
cycler==0.11.0
environs==8.0.0
fonttools==4.38.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
kombu==5.2.4
MarkupSafe==2.1.2
marshmallow==3.18.0
matplotlib==3.6.3
numpy==1.24.1
packaging==21.3
pdfkit==1.0.0
peewee==3.14.10
peewee-migrate==1.4.8
Pillow==9.4.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-telegram-bot==20.1
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
redis==4.5.4
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tornado==6.2
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.6
